chore: consolidate loki twin pairs and cluster test files — issue #1449
- Merge authz.loki.test.ts → authz.test.ts (21 edge-case tests)
- Merge decree-parser.loki.test.ts → decree-parser.test.ts (17 tests)
- Merge pull.route.loki.test.ts → pull.route.test.ts (3 IDOR invariant tests)
- Delete push.idor.loki.test.ts (fully duplicated by push.route.test.ts)
- Collapse 5 use-cloud-sync issue files → use-cloud-sync-regressions.test.ts
- Merge trial/trial-store.test.ts → lib/trial-store.test.ts (boundary + TTL tests)
- Merge anon-settings-1345-loki.test.tsx → anon-settings-1345.test.tsx (8 tests)

All 2358 tests pass. tsc clean. Next.js build clean (45 routes).
