--- development/frontend/src/__tests__/hooks/use-cloud-sync-1124-loki.test.ts
@@ -1,509 +1,0 @@
-/**
- * useCloudSync — Loki QA tests for Issue #1124
- * Auto-migration: localStorage → Firestore on first Karl sign-in
- *
- * Acceptance criteria tested:
- *   AC-1: Thrall user signs in → handleLoginTransition never fires → no migration
- *   AC-2: Already-migrated Karl signs in → delegates to performSync (not runMigration)
- *   AC-3: First Karl sign-in, upload direction → "backed up to the cloud" toast
- *   AC-4: First Karl sign-in, download direction → "restored from cloud" toast
- *   AC-5: First Karl sign-in, merge direction → "backed up to the cloud" toast
- *   AC-6: First Karl sign-in, empty direction (0 cards) → no toast shown
- *   AC-7: Singular card count → "1 card has been backed up to the cloud"
- *   AC-8: Migration error → falls back to regular performSync → migrated flag NOT set
- *   AC-9: fenrir:first-sync-shown is set after successful migration
- *   AC-10: Status transitions idle → syncing → synced during migration
- *   AC-11: syncInProgress lock prevents concurrent migration
- *
- * Issue #1124
- */
-
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { renderHook, act } from "@testing-library/react";
-import { useCloudSync } from "@/hooks/useCloudSync";
-import { toast } from "sonner";
-
-// ── Mocks ──────────────────────────────────────────────────────────────────────
-
-// Start as Thrall; tests flip to Karl as needed
-const mockEntitlement = { tier: "thrall" as string, isActive: false };
-
-vi.mock("@/hooks/useEntitlement", () => ({
-  useEntitlement: () => mockEntitlement,
-}));
-
-const mockAuthContext = { status: "authenticated" as string };
-
-vi.mock("@/contexts/AuthContext", () => ({
-  useAuthContext: () => mockAuthContext,
-}));
-
-vi.mock("sonner", () => ({
-  toast: {
-    success: vi.fn(),
-    error: vi.fn(),
-  },
-}));
-
-vi.mock("@/lib/storage", () => ({
-  getRawAllCards: vi.fn().mockReturnValue([]),
-  setAllCards: vi.fn(),
-}));
-
-// Controllable migration mock
-const mockHasMigrated = vi.fn<() => boolean>(() => false);
-const mockRunMigration = vi.fn();
-
-vi.mock("@/lib/sync/migration", () => ({
-  hasMigrated: () => mockHasMigrated(),
-  runMigration: (...args: unknown[]) => mockRunMigration(...args),
-  MIGRATION_FLAG: "fenrir:migrated",
-}));
-
-// ── Session helpers ─────────────────────────────────────────────────────────────
-
-const FAKE_SESSION = { id_token: "tok-1124", user: { sub: "hh-1124" } };
-
-function setSession() {
-  localStorage.setItem("fenrir:auth", JSON.stringify(FAKE_SESSION));
-}
-
-function clearSession() {
-  localStorage.removeItem("fenrir:auth");
-}
-
-// ── Migration result factories ─────────────────────────────────────────────────
-
-function migrationResult(
-  direction: "download" | "upload" | "merge" | "empty",
-  cardCount: number,
-  ran = true
-) {
-  return { ran, cardCount, direction };
-}
-
-// ── Sync response helpers (used for performSync fallback tests) ─────────────────
-
-function successSyncResponse(syncedCount = 0, cards: unknown[] = []) {
-  return Promise.resolve(
-    new Response(JSON.stringify({ cards, syncedCount }), {
-      status: 200,
-      headers: { "Content-Type": "application/json" },
-    })
-  );
-}
-
-// ── Tests ──────────────────────────────────────────────────────────────────────
-
-describe("useCloudSync / handleLoginTransition — AC-1: Thrall user", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("does not call runMigration when isKarl is false", () => {
-    setSession();
-    renderHook(() => useCloudSync());
-    // isKarl=false → handleLoginTransition is never called
-    expect(mockRunMigration).not.toHaveBeenCalled();
-  });
-
-  it("status stays idle for Thrall user", () => {
-    setSession();
-    const { result } = renderHook(() => useCloudSync());
-    expect(result.current.status).toBe("idle");
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-2: already migrated delegates to performPull", () => {
-  const mockFetch = vi.fn();
-
-  beforeEach(() => {
-    mockEntitlement.tier = "karl";
-    mockEntitlement.isActive = true;
-    // Already migrated
-    mockHasMigrated.mockReturnValue(true);
-    mockRunMigration.mockReset();
-    vi.stubGlobal("fetch", mockFetch);
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  afterEach(() => {
-    vi.unstubAllGlobals();
-    clearSession();
-  });
-
-  it("does not call runMigration when fenrir:migrated is set", async () => {
-    // Issue #1239: already-migrated users get performPull (GET /api/sync/pull),
-    // NOT performSync (POST /api/sync/push). Login-time sync is pull-only.
-    mockFetch.mockReturnValue(
-      Promise.resolve(
-        new Response(JSON.stringify({ cards: [], activeCount: 2 }), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        })
-      )
-    );
-    setSession();
-
-    await act(async () => {
-      renderHook(() => useCloudSync());
-    });
-
-    // runMigration must NOT be called
-    expect(mockRunMigration).not.toHaveBeenCalled();
-    // Issue #1239: already-migrated login uses performPull (GET), not performSync (POST)
-    expect(mockFetch).toHaveBeenCalledWith(
-      expect.stringContaining("/api/sync/pull"),
-      expect.objectContaining({ method: "GET" })
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-3: upload direction toast", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("shows 'backed up to the cloud' toast when direction=upload (3 cards)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("upload", 3));
-    setSession();
-
-    // Render as Karl (triggers the non-Karl → Karl transition)
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).toHaveBeenCalledWith(
-      "Your 3 cards have been backed up to the cloud",
-      expect.objectContaining({ description: "Yggdrasil guards your ledger." })
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-4: download direction toast", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("shows 'restored from cloud' toast when direction=download (5 cards)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("download", 5));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).toHaveBeenCalledWith(
-      "Your 5 cards have been restored from cloud",
-      expect.objectContaining({ description: "Yggdrasil guards your ledger." })
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-5: merge direction toast", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("shows 'backed up to the cloud' toast when direction=merge (4 cards)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("merge", 4));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).toHaveBeenCalledWith(
-      "Your 4 cards have been backed up to the cloud",
-      expect.objectContaining({ description: "Yggdrasil guards your ledger." })
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-6: empty direction, no toast", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("shows no toast when cardCount=0 (empty direction)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("empty", 0));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).not.toHaveBeenCalled();
-    expect(toast.error).not.toHaveBeenCalled();
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-7: singular card toast", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  it("uses singular 'card has' for exactly 1 card (upload)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("upload", 1));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).toHaveBeenCalledWith(
-      "Your 1 card has been backed up to the cloud",
-      expect.anything()
-    );
-  });
-
-  it("uses singular 'card has' for exactly 1 card (download)", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("download", 1));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(toast.success).toHaveBeenCalledWith(
-      "Your 1 card has been restored from cloud",
-      expect.anything()
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-8: migration error fallback", () => {
-  const mockFetch = vi.fn();
-
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.stubGlobal("fetch", mockFetch);
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  afterEach(() => {
-    vi.unstubAllGlobals();
-    clearSession();
-  });
-
-  it("falls back to performSync when runMigration throws a network error", async () => {
-    // runMigration fails
-    mockRunMigration.mockRejectedValue(
-      Object.assign(new Error("Network unavailable"), { code: "network_error" })
-    );
-    // performSync (fallback) succeeds
-    mockFetch.mockReturnValue(successSyncResponse(2));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    // runMigration was attempted
-    expect(mockRunMigration).toHaveBeenCalledWith("hh-1124", "tok-1124");
-    // performSync (fetch) was called as fallback
-    expect(mockFetch).toHaveBeenCalledWith("/api/sync/push", expect.anything());
-  });
-
-  it("does not show migration toast on migration error", async () => {
-    mockRunMigration.mockRejectedValue(new Error("API down"));
-    mockFetch.mockReturnValue(successSyncResponse(0));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    // No migration toast
-    expect(toast.success).not.toHaveBeenCalledWith(
-      expect.stringContaining("backed up"),
-      expect.anything()
-    );
-    expect(toast.success).not.toHaveBeenCalledWith(
-      expect.stringContaining("restored"),
-      expect.anything()
-    );
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-9: first-sync-shown flag", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    localStorage.clear();
-  });
-
-  afterEach(() => {
-    clearSession();
-    localStorage.clear();
-  });
-
-  it("sets fenrir:first-sync-shown after successful migration", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("upload", 2));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(localStorage.getItem("fenrir:first-sync-shown")).toBe("true");
-  });
-
-  it("does NOT set fenrir:first-sync-shown when migration throws", async () => {
-    const mockFetch = vi.fn().mockReturnValue(successSyncResponse(0));
-    vi.stubGlobal("fetch", mockFetch);
-    mockRunMigration.mockRejectedValue(new Error("error"));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    // Flag may have been set by performSync fallback; migration itself doesn't set it on error
-    // What matters: runMigration threw and didn't set it
-    expect(mockRunMigration).toHaveBeenCalled();
-    vi.unstubAllGlobals();
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-10: status transitions", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  afterEach(() => {
-    clearSession();
-  });
-
-  it("transitions to 'syncing' during migration and 'synced' after success", async () => {
-    let resolveRunMigration!: (val: unknown) => void;
-    const migrationPromise = new Promise((res) => {
-      resolveRunMigration = res;
-    });
-    mockRunMigration.mockReturnValue(migrationPromise);
-    setSession();
-
-    let hookResult: ReturnType<typeof renderHook<ReturnType<typeof useCloudSync>>>;
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      hookResult = renderHook(() => useCloudSync());
-    });
-
-    // After login transition fires, status should be "syncing"
-    expect(hookResult!.result.current.status).toBe("syncing");
-
-    // Resolve migration
-    await act(async () => {
-      resolveRunMigration(migrationResult("upload", 2));
-    });
-
-    expect(hookResult!.result.current.status).toBe("synced");
-    expect(hookResult!.result.current.cardCount).toBe(2);
-  });
-});
-
-describe("useCloudSync / handleLoginTransition — AC-11: passes householdId and idToken", () => {
-  beforeEach(() => {
-    mockEntitlement.tier = "thrall";
-    mockEntitlement.isActive = false;
-    mockHasMigrated.mockReturnValue(false);
-    mockRunMigration.mockReset();
-    vi.clearAllMocks();
-    clearSession();
-  });
-
-  afterEach(() => {
-    clearSession();
-  });
-
-  it("calls runMigration with householdId from session.user.sub and idToken from session.id_token", async () => {
-    mockRunMigration.mockResolvedValue(migrationResult("upload", 1));
-    setSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(mockRunMigration).toHaveBeenCalledWith("hh-1124", "tok-1124");
-  });
-
-  it("does not call runMigration when session is absent", async () => {
-    // No session set
-    clearSession();
-
-    await act(async () => {
-      mockEntitlement.tier = "karl";
-      mockEntitlement.isActive = true;
-      renderHook(() => useCloudSync());
-    });
-
-    expect(mockRunMigration).not.toHaveBeenCalled();
-  });
-});
