+++ development/frontend/src/__tests__/lib/trial-store.test.ts
@@ -2,6 +2,7 @@
  * Unit tests for kv/trial-store.ts — Redis trial CRUD operations.
  *
  * Mocks ioredis via redis-client to test get/set paths and error handling.
+ * Includes boundary condition tests merged from trial/trial-store.test.ts (issue #944).
  */
 
 import { describe, it, expect, beforeEach, vi } from "vitest";
@@ -17,10 +18,15 @@ vi.mock("@/lib/kv/redis-client", () => ({
   getRedisClient: () => mockRedis,
 }));
 
+vi.mock("@/lib/logger", () => ({
+  log: { debug: vi.fn(), info: vi.fn(), warn: vi.fn(), error: vi.fn() },
+}));
+
 // ── Import after mock ────────────────────────────────────────────────────
 
 import { getTrial, initTrial, markTrialConverted, computeTrialStatus } from "@/lib/kv/trial-store";
 import type { StoredTrial } from "@/lib/kv/trial-store";
+import { TRIAL_DURATION_DAYS } from "@/lib/trial-utils";
 
 // ── Test data ────────────────────────────────────────────────────────────
 
@@ -158,5 +164,80 @@ describe("trial-store", () => {
       expect(result.status).toBe("active");
       expect(result.remainingDays).toBeGreaterThan(0);
     });
+  });
+});
+
+// ---------------------------------------------------------------------------
+// Loki boundary condition tests (merged from trial/trial-store.test.ts — issue #944)
+// ---------------------------------------------------------------------------
+
+function daysAgo(n: number): string {
+  const d = new Date(Date.now() - n * 24 * 60 * 60 * 1000);
+  return d.toISOString();
+}
+
+const VALID_FINGERPRINT_B = "b".repeat(64);
+
+describe("computeTrialStatus — boundary conditions (#944 AC2)", () => {
+  it("returns status:none when trial is null (no record in Redis)", () => {
+    const result = computeTrialStatus(null);
+    expect(result.status).toBe("none");
+    expect(result.remainingDays).toBe(0);
+  });
+
+  it(`returns active with 1 remaining day when exactly ${TRIAL_DURATION_DAYS - 1} days have elapsed`, () => {
+    const startDate = daysAgo(TRIAL_DURATION_DAYS - 1); // day 29
+    const result = computeTrialStatus({ startDate });
+    expect(result.status).toBe("active");
+    expect(result.remainingDays).toBe(1);
+  });
+
+  it(`returns expired with 0 days when exactly ${TRIAL_DURATION_DAYS} days have elapsed`, () => {
+    const startDate = daysAgo(TRIAL_DURATION_DAYS); // day 30 — trial is over
+    const result = computeTrialStatus({ startDate });
+    expect(result.status).toBe("expired");
+    expect(result.remainingDays).toBe(0);
+  });
+
+  it("returns converted (overriding expiry) when convertedDate is set (#944 AC3)", () => {
+    const startDate = daysAgo(TRIAL_DURATION_DAYS + 5); // well past expiry
+    const convertedDate = daysAgo(TRIAL_DURATION_DAYS - 20);
+    const result = computeTrialStatus({ startDate, convertedDate });
+    expect(result.status).toBe("converted");
+    expect(result.remainingDays).toBe(0);
+    expect(result.convertedDate).toBe(convertedDate);
+  });
+});
+
+describe("initTrial — Redis key format and TTL (#944 AC1)", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it("writes trial under key `trial:{fingerprint}` with TTL 5184000 s (60 days)", async () => {
+    mockRedis.get.mockResolvedValueOnce(null); // no existing trial
+    mockRedis.set.mockResolvedValueOnce("OK");
+
+    await initTrial(VALID_FINGERPRINT_B);
+
+    expect(mockRedis.set).toHaveBeenCalledOnce();
+    const [key, , exFlag, ttl] = mockRedis.set.mock.calls[0] as [string, string, string, number];
+    expect(key).toBe(`trial:${VALID_FINGERPRINT_B}`);
+    expect(exFlag).toBe("EX");
+    // 60 days × 24 h × 60 min × 60 s = 5 184 000 s
+    expect(ttl).toBe(60 * 24 * 60 * 60);
+  });
+});
+
+describe("getTrial — malformed JSON error handling", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  it("returns null (not throws) when Redis contains malformed JSON", async () => {
+    mockRedis.get.mockResolvedValueOnce("not-valid-json{{{");
+
+    const result = await getTrial(VALID_FINGERPRINT_B);
+    expect(result).toBeNull();
   });
 });
